refactor: extract JavaPrivateKey and JavaPublicKey into separate files
