+++ projects/core/src/main/scala/crypto4s/JavaPrivateKey.scala
@@ -1,0 +1,10 @@
+package crypto4s
+
+import java.security.PrivateKey as JPrivateKey
+
+private[crypto4s] case class JavaPrivateKey[Alg](
+  delegate: JPrivateKey
+) extends PrivateKey[Alg] {
+
+  override def asJava: JPrivateKey = delegate
+}
