+++ projects/core/src/main/scala/crypto4s/JavaPublicKey.scala
@@ -1,0 +1,10 @@
+package crypto4s
+
+import java.security.PublicKey as JPublicKey
+
+private[crypto4s] case class JavaPublicKey[Alg](
+  delegate: JPublicKey
+) extends PublicKey[Alg] {
+
+  override def asJava: JPublicKey = delegate
+}
